rhombus_data_infer/requirements.txt
asgiref==3.7.2
Django==4.2.11
django-cors-headers==4.3.1
django-filter==24.1
djangorestframework==3.14.0
importlib_metadata==7.0.2
Markdown==3.6
numpy==1.26.4
pandas==2.2.1
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
zipp==3.18.0
